+++ core/src/main/scala/com/ccadllc/cedi/dtrace/SpanId.scala
@@ -19,8 +19,6 @@ import java.util.UUID
 
 import scala.language.higherKinds
 import scala.util.Random
-import scala.util.control.NonFatal
-import scala.util.matching.Regex
 
 import cats.effect.Sync
 import cats.implicits._
@@ -41,13 +39,6 @@ final case class SpanId(traceId: UUID, parentSpanId: Long, spanId: Long) {
   val root: Boolean = parentSpanId == spanId
 
   /**
-   * Converts the span to a `Money`-compliant HTTP Header.
-   * @return header - the `Money`-compliant header consisting of
-   *   `X-MoneyTrace=trace-id=<trace id UUID>;parent-id=<long integer>;span-id=<long integer>`
-   */
-  val toHeader: String = s"${SpanId.TraceIdHeader}=$traceId;${SpanId.ParentIdHeader}=$parentSpanId;${SpanId.SpanIdHeader}=$spanId"
-
-  /**
    * Creates a new child span identifier from this instance, with the new instance's `parentSpanId` equal to this instance's `spanId` and a new value generated
    * for the new instance's `spanId`.
    * @return newSpanIdDescription - an effectful description of a new [[SpanId]].
@@ -61,18 +52,6 @@ final case class SpanId(traceId: UUID, parentSpanId: Long, spanId: Long) {
  * Companion object of the `SpanId` datatype.  Provides smart constructors and commonly used constants.
  */
 object SpanId {
-  /** The `Money` compliant HTTP header name. */
-  final val HeaderName: String = "X-MoneyTrace"
-  /** The `Money` compliant HTTP header trace GUID component value identifier. */
-  final val TraceIdHeader: String = "trace-id"
-  /** The `Money` compliant HTTP header Parent Span ID component value identifier. */
-  final val ParentIdHeader: String = "parent-id"
-  /** The `Money` compliant HTTP header Span ID component value identifier. */
-  final val SpanIdHeader: String = "span-id"
-
-  /* Used to validate / parse `Money` compliant HTTP header into a [[SpanId]] instance. */
-  final val HeaderRegex: Regex = s"$TraceIdHeader=([0-9a-f]{8}(-[0-9a-f]{4}){3}-[0-9a-fA-F]{12});$ParentIdHeader=([\\-0-9]+);$SpanIdHeader=([\\-0-9]+)".r
-
   /**
    * Creates a root [[SpanId]] from stratch in an effectful program `F[A]`.
    *
@@ -82,24 +61,6 @@ object SpanId {
     traceId <- F.delay(UUID.randomUUID)
     parentChildId <- nextSpanIdValue
   } yield SpanId(traceId, parentChildId, parentChildId)
-
-  /**
-   * Creates an instance of a [[SpanId]] from the given header name and value, if the
-   * header name matches the `Money` compliant header name of `X-MoneyTrace` and the header
-   * value is in the proper format; otherwise, an error is returned.
-   * @return newSpanIdOrError - a [[SpanId]] if successful; otherwise, an error message is returned.
-   */
-  def fromHeader(headerName: String, headerValue: String): Either[String, SpanId] =
-    if (headerName == HeaderName) fromHeaderValue(headerValue) else Left(s"Header name $headerName is not a Money-compliant trace header")
-
-  def fromHeaderValue(headerValue: String): Either[String, SpanId] = headerValue match {
-    case HeaderRegex(traceId, _, parentId, spanId) =>
-      try Right(SpanId(UUID.fromString(traceId), parentId.toLong, spanId.toLong))
-      catch {
-        case NonFatal(t) => Left(s"Could not parse $headerValue into a SpanId: ${t.getMessage}")
-      }
-    case _ => Left(s"Could not parse $headerValue into a SpanId")
-  }
 
   private[dtrace] def nextSpanIdValue[F[_]](implicit F: Sync[F]): F[Long] = F.delay(Random.nextLong)