+++ CHANGELOG.md
@@ -1,6 +1,12 @@
+
 1.3.0
 =====
- - Implement ConcurrentEffect typeclass
+ - Implement ConcurrentEffect and Timer typeclass
+ - Fixed issue with Effect typeclass runAsync function implementation
+   causing loss of trace context.
+ - Added money, xb3, and http4s modules, with support for Money,
+ - and X-B3 headers and generation and extraction of trace information to/from
+ - headers for the http4s library.
 
 1.1.0
 =====